Update `assert_within_limits` on the `Prior` class
This update makes the `assert_within_limits` method of the `Prior` class play nicely with JAX.  As the `value` passed into the function will be a JAX traced array a `jax.lax.cond` must be used in place of a traditional `if` block.

As one branch will `raise` and exception it must be wrapped in a `jax.debug.callback()` to ensure it does not evaluate until the `jit`ed function exits.

Also the `jax.numpy.logical_*` are used in the condition definition to make it work on the traced array as expected.

Side note: the `if jax_wrapper.use_jax:` block does not need this kind of treatment as it is a static value when the function is `jit`ed and the compiler knows what branch to take.